\name{Findwords}
\alias{hello}
\title{Hello, World!}
\usage{
francaise_english("")
deutsch_english("")
}
\description{
This package was designed to help create Anki decks for language learning. You type in a word in your target language, and it will open 3 website in your browser.

\enumerate{
\item Audio fioles for pronounciation via "forvo.com""
\item Google image result result for visual representation
\item Translation
}

Supported languages currenlty : French and German

Looking to devleope this futher, suggestions always welcome.
}
\examples{
francaise_english("fromage")

**browser will open results of french word**

deutsch_english("käse")

**browser will open results of german word**
}
